Added some audio scripts
Yesterday was the day I destroyed everything. So right now I'm adding some of the first new versions of the scripts. This includes a new config_parser, (which reads a config file and makes a dictionary), an audio_metadata_print script (which just prints ID3 tags right now) and the beginnings of a beautifully created metadata transfer script.

~ Love, Ted M.