Added ripping script
Today I completed a lot of work in the way of creating a new working audio CD ripper. And it works! There are still quite a few things left to be dealt with, but many of the major features work perfectly. You can rip a CD to FLAC as long as you have a configuration file, a genres file, (both in ~/.config) and a properly named album artwork file on your hard drive. The application of metadata and album art work perfectly when all of these requisites areavailable.

~ Love, Ted M.