Updated audio ripper and README
Over the past few days, I have updated the ripper script to be much more effective. The new config file has more options, and now album art seems to work just fine. It may be a little messy though, so I will have to comment it further in the future.

The README.adoc is now slightly more specific as to the contents of the repo

~ Love, Ted M.